xbee sending almost works
however, still sending some garbage values when not sending. Data sends and is received correctly, just there is extra garbage being printed.
This may have something to do with the board I'm using right now, which is the USB explorer, i read something about it being a problem because the RSSI pin is connected to another pull-up resistor or what not. We'll see what happens when we switch back to the mbed standard breakout board.